+++ src/charts/fixtures/pie-chart-custom-colors.js
@@ -1,0 +1,35 @@
+export const data = [
+  {
+    name: 'Light Light Grey',
+    price: 'EUR 5.282,90',
+    y: 10,
+  },
+  {
+    name: 'Light Grey',
+    price: 'EUR 5.282,90',
+    y: 20,
+  },
+  {
+    name: 'Grey',
+    price: 'EUR 5.282,90',
+    y: 30,
+  },
+  {
+    name: 'Dark Grey',
+    price: 'EUR 5.282,90',
+    y: 40,
+  },
+  {
+    name: 'Dark Dark Grey',
+    price: 'EUR 5.282,90',
+    y: 50,
+  }
+];
+
+export const colors = [
+  '#d8d8d8',
+  '#b1b1b1',
+  '#7e7e7e',
+  '#474747',
+  '#1a1a1a'
+];